+++ code/engine.vc2008/xrEngine/DiscordRichPresense.cpp
@@ -1,8 +1,10 @@
 #include "stdafx.h"
+#include "../xrCore/StringsUtils.h"
 #include "DiscordRichPresense.h"
 #include "DiscordRichPresense/discord_register.h"
 #include "DiscordRichPresense/discord_rpc.h"
 #include "x_ray.h"
+#include "IGame_Level.h"
 
 ENGINE_API xrDiscordPresense g_discord;
 
@@ -30,7 +32,7 @@ void xrDiscordPresense::SetStatus(StatusId status)
 	if (!bInitialize) return;
 
 	DiscordRichPresence presenseInfo;
-	ZeroMemory(&presenseInfo, sizeof(presenseInfo));
+	memset(&presenseInfo, 0, sizeof(presenseInfo));
 
 	StatusId realStatus = status;
 	if (status == StatusId::In_Game)
@@ -56,30 +58,34 @@ void xrDiscordPresense::SetStatus(StatusId status)
 	}
 
 	presenseInfo.largeImageText = "Zone Awaits...";
+	presenseInfo.details = "Level: ";
+
+	bool bOnLevel = true;
 
 	switch (realStatus)
 	{
-	case StatusId::In_Game:
-		presenseInfo.details		= "In Game";
-		presenseInfo.largeImageKey	= "zaton";
-		break;
-	case StatusId::Zaton:
-		presenseInfo.details		= "In Game (Zaton)";
-		presenseInfo.largeImageKey  = "zaton";
-		break;
-	case StatusId::Upiter:
-		presenseInfo.details		= "In Game (Jupiter)";
-		presenseInfo.largeImageKey  = "ypiter";
-		break;
-	case StatusId::Pripyat:
-		presenseInfo.details		= "In Game (Pripyat)";
-		presenseInfo.largeImageKey  = "pripyat";
-		break;
-	default:
+	case StatusId::In_Game:	presenseInfo.largeImageKey	= "zaton"; break;
+	case StatusId::Zaton:	presenseInfo.largeImageKey  = "zaton"; break;
+	case StatusId::Upiter:	presenseInfo.largeImageKey  = "ypiter"; break;
+	case StatusId::Pripyat:	presenseInfo.largeImageKey  = "pripyat"; break;
 	case StatusId::Menu:
-		presenseInfo.details		= "In Menu";
+	default:
+		presenseInfo.details		= "Main Menu";
 		presenseInfo.largeImageKey	= "menu_final";
+		bOnLevel = false;
 		break;
+	}
+
+	// Add levelname to details
+	if (bOnLevel && g_pGameLevel && g_pGameLevel->name_translated().c_str())
+	{
+		string64 LevelName;
+		strconcat(sizeof(LevelName), LevelName, "Level: ", ConvertToUTF8(g_pGameLevel->name_translated().c_str()));
+		presenseInfo.details = std::move(LevelName);
+	}
+	else if (!g_pGameLevel)
+	{
+		presenseInfo.details = "Loading...";
 	}
 
 	Discord_UpdatePresence(&presenseInfo);
